Character Select is Pretty Much done and health bars work!
I Sliced up the Meldin Sprites and added him the the char select and also put all the lockin voicelines in. The Health bars work and reset properly but more work is needed in the fighting scene. like the super and block bars and making sure the character can't move during round transitions.